+++ packages/salesforcedx-vscode-core/src/commands/forceSourceDeployManifest.ts
@@ -4,48 +4,19 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import { ContinueResponse } from '@salesforce/salesforcedx-utils-vscode';
-import { ComponentSet } from '@salesforce/source-deploy-retrieve';
-import { join } from 'path';
 import * as vscode from 'vscode';
 import { channelService } from '../channels';
 import { TimestampConflictChecker } from '../commands/util/postconditionCheckers';
 import { getConflictMessagesFor } from '../conflict/messages';
 import { nls } from '../messages';
 import { notificationService } from '../notifications';
-import { SfdxPackageDirectories } from '../sfdxProject';
 import { telemetryService } from '../telemetry';
-import { workspaceUtils } from '../util';
-import { DeployExecutor } from './baseDeployRetrieve';
+import { LibrarySourceDeployManifestExecutor } from './librarySourceDeployManifestExecutor';
 import {
-  ConflictDetectionMessages,
   FilePathGatherer,
   SfdxCommandlet,
   SfdxWorkspaceChecker
 } from './util';
-
-export class LibrarySourceDeployManifestExecutor extends DeployExecutor<
-  string
-> {
-  constructor() {
-    super(
-      nls.localize('force_source_deploy_text'),
-      'force_source_deploy_with_manifest_beta'
-    );
-  }
-
-  protected async getComponents(
-    response: ContinueResponse<string>
-  ): Promise<ComponentSet> {
-    const packageDirs = await SfdxPackageDirectories.getPackageDirectoryPaths();
-    return ComponentSet.fromManifest({
-      manifestPath: response.data,
-      resolveSourcePaths: packageDirs.map(dir =>
-        join(workspaceUtils.getRootWorkspacePath(), dir)
-      )
-    });
-  }
-}
 
 export async function forceSourceDeployManifest(manifestUri: vscode.Uri) {
   if (!manifestUri) {
