+++ packages/salesforcedx-vscode-core/src/commands/forceSourceRetrieveManifest.ts
@@ -4,53 +4,17 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import {
-  Command,
-  SfdxCommandBuilder
-} from '@salesforce/salesforcedx-utils-vscode';
-import { ContinueResponse } from '@salesforce/salesforcedx-utils-vscode';
-import { ComponentSet } from '@salesforce/source-deploy-retrieve';
-import { join } from 'path';
 import * as vscode from 'vscode';
 import { channelService } from '../channels';
 import { nls } from '../messages';
 import { notificationService } from '../notifications';
-import { sfdxCoreSettings } from '../settings';
-import { SfdxPackageDirectories } from '../sfdxProject';
 import { telemetryService } from '../telemetry';
-import { workspaceUtils } from '../util';
-import { RetrieveExecutor } from './baseDeployRetrieve';
+import { LibrarySourceRetrieveManifestExecutor } from './librarySourceRetrieveManifestExecutor';
 import {
   FilePathGatherer,
   SfdxCommandlet,
-  SfdxCommandletExecutor,
   SfdxWorkspaceChecker
 } from './util';
-
-export class LibrarySourceRetrieveManifestExecutor extends RetrieveExecutor<
-  string
-> {
-  constructor() {
-    super(
-      nls.localize('force_source_retrieve_text'),
-      'force_source_retrieve_with_manifest_beta'
-    );
-  }
-
-  protected async getComponents(
-    response: ContinueResponse<string>
-  ): Promise<ComponentSet> {
-    const packageDirs = await SfdxPackageDirectories.getPackageDirectoryPaths();
-
-    return ComponentSet.fromManifest({
-      manifestPath: response.data,
-      resolveSourcePaths: packageDirs.map(relativeDir =>
-        join(workspaceUtils.getRootWorkspacePath(), relativeDir)
-      ),
-      forceAddWildcards: true
-    });
-  }
-}
 
 export async function forceSourceRetrieveManifest(explorerPath: vscode.Uri) {
   if (!explorerPath) {
