+++ packages/salesforcedx-vscode-core/test/jest/commands/retrieveExecutor.test.ts
@@ -11,7 +11,7 @@ import {
 } from '@salesforce/salesforcedx-utils-vscode';
 import { ComponentSet } from '@salesforce/source-deploy-retrieve';
 import * as fs from 'fs';
-import { RetrieveExecutor } from '../../../src/commands/baseDeployRetrieve';
+import { RetrieveExecutor } from '../../../src/commands/retrieveExecutor';
 import { WorkspaceContext } from '../../../src/context/workspaceContext';
 
 jest.mock('../../../src/sfdxProject/sfdxProjectConfig');
